+++ scripts/buckeye_experiments/baseline_eval.sh
@@ -1,11 +1,11 @@
 #!/bin/bash
 
-#SBATCH -c 12
+#SBATCH -c 8
 #SBATCH --mem=8GB
 #SBATCH -p gpu-preempt
 #SBATCH -G 1
 #SBATCH --time 09:00:00
-#SBATCH -o baseline_eval.out
+#SBATCH -o %j_baseline_eval.out
 #SBATCH --mail-type END
 
 
@@ -16,10 +16,10 @@ EVAL_RESULTS_CSV=data/evaluation_results/aggregate_metrics/baseline_eval.csv
 DETAILED_RESULTS_DIR=data/evaluation_results/detailed_predictions
 DATA_DIR=data/buckeye
 
-module load miniconda/22.11.1-1
-conda activate ./env
+module load conda/latest
+conda activate ./env_cuda124
 
-multipa-evaluate --hf_models ctaguchi/wav2vec2-large-xlsr-japlmthufielta-ipa-plus-2000 \
+multipa-evaluate --hf_models ctaguchi/wav2vec2-large-xlsr-japlmthufielta-ipa-plus-2000 facebook/wav2vec2-lv-60-espeak-cv-ft \ 
  --local_models data/models/hyperparam_tuning_1/wav2vec2-large-xlsr-buckeye-ipa \
  --eval_out $EVAL_RESULTS_CSV \
  --verbose_results_dir $DETAILED_RESULTS_DIR \